requirements.txt
### Add more library
### jsonpatch, jsonpointer cause json_helper.py required
### pymssql-py38 cause pymssql is deprecated
#add new library 23-09-2020 : keyboard: pip install keyboard

openpyxl==2.6.3
PyJWT==1.7.1
pylint==1.9.5
pymssql==2.1.4
PyYAML==5.1.2
requests==2.22.0
robotframework==3.1.2
robotframework-databaselibrary==1.2.4
robotframework-excellibrary==0.0.2
robotframework-jsonlibrary==0.3.1
robotframework-requests==0.5.0
robotframework-seleniumlibrary==4.0.0
selenium==3.141.0
robotframework-debuglibrary==1.1.4   
jsonpatch==1.24
jsonpointer==2.0
pymssql-py38==2.1.4  
keyboard==0.13.5

